리팩토링: ExternalAiClientImpl 책임별 클라이언트 분리
1개 클래스(293줄)에 GPT/YOLO/LAMA/S3 4개 외부 시스템이 섞여있던 구조를
시스템별 클라이언트로 분리하고, 기존 클래스는 thin composite로 슬림화.

- GptVisionClient: GPT-4 Vision 분석/비교 + 응답 파싱
- YoloDetectionClient: 어지러운 영역 감지
- LamaInpaintClient: 인페인팅 + 전용 RestTemplate(multipart)
- S3ImageStore: AI 파이프라인 전용 S3 헬퍼(다운로드/Base64/업로드)
- ExternalAiClientImpl: 293줄 → 50줄 thin composite로 위임

도메인 인터페이스(ExternalAiClient)는 미변경 → 도메인 영향 없음.

부수 개선:
- requestCompareAnalysis의 catch(Exception)이 OpenAI 429를 묵살하던
  사일런트 버그 수정 → AI_RATE_LIMIT_EXCEEDED로 일관 처리
- 사용되지 않던 @Value 필드 3개 제거(openai.api.url, openai.api.url.image, openai.api.key)
